+++ src/types/markdown-it-katex.d.ts
@@ -1,0 +1,5 @@
+declare module 'markdown-it-katex' {
+    import MarkdownIt from 'markdown-it';
+    const mkdnKatex: MarkdownIt.PluginSimple | MarkdownIt.PluginWithOptions;
+    export default mkdnKatex;
+}